Disable an infinite loop check for while statements
Error checking for infinite loops was added in PR #1213, but the
implementation is problematic:
 - it only applies to while statements
 - it does not handle 'return' or 'break'

Besides, it's not obvious to me that a program containing an infinite
loop is actually an error.

So, this commit disables that check for the near term.
See also issue #6532.

Note that there still remains one check of inifinite loops that
prints out `A while loop with a constant condition`. If we choose
to remove that, it will be in a different PR.