Improving --log support
chpl --log user-facing changes:
  * Removes --log= in favor of --log-pass=.
    --log=<letter> is no longer available. Instead, we have
    --log (to enable logging) and --log-pass=<letter>.
    Additionally, --log-pass can take in a pass name (not just a short
    name). Lastly, --log-pass=<something> activates --log,
    so when selecting a pass, only --log-pass=something is required.
  * Adds --log-node option, to select AstDumpToNode instead of
    AstDump.
  * makes --log-module=, --log-pass, --log-dir also activate --log,
    so that both need not be specified.
  * removes the log directory before outputting any AST dumps

Besides these changes, updated logging code to work with string pass
names instead of 1-char pass IDs. This allows the 1-char pass ID to be
omitted and for the pass to be still logged when logging all passes.

To facilitate that, log.h now defines LOG_NEVER, which means a pass
should not be logged with --log (but it can stil be selected directly
with --log-pass=<name>) and LOG_NO_SHORT, which means that that pass does
not have a 1-character short name.

While there I removed all of the 1-character short names that do not seem
mnemonic to me, since a short name is no longer necessary to get AST
dumps.