Rewrite this '=>' by just replacing the LHS with the RHS
Reason: using '=>' had the benefit of creating a privatized copy of
the array such that every locale had its own local copy of the
array and could access it within a local block.  With 'ref's this
is not the case.  But the renaming here was merely a convenience
anyway, so let's just use the less convenient name.