Implement reindex() of an empty domain/array
In some work earlier today, I was running into an issue that, on the
way home, I realized was likely due to the .reindex() operation not
working properly when applied to an empty domain (or a domain with an
empty range).  Reproducing this in a small test (added here as
emptyReindex.chpl), I verified that it did not work.  Changing the
reindex operation to make it work was essentially a five-minute fix to
detect and handle the empty range case.  This also simplifies some
logic from the previous implementation by handling a dimension at a
time rather than working at the tuple-of-indices level.