Partial fix of issue encountered when converting a record to use initializers
The assertion that concrete records with initializers did not need a _defaultOf
had missed handling `out` arguments of that type.  This change rectifies that
oversight, by updating the section which inserts a PRIM_INIT call for out
arguments to instead use an init call when the type is a concrete record that
defines initializers.