Add a boiled down version of an error I am encountering with autoCopy & init()
When attempting to convert test/types/records/noakes/return.chpl to using an
initializer, I encountered an error when we returned a record that defined an
initializer to an instance that already knew its type.  That error is due to
our implementation of copy initializers, I believe.  We don't validate our
autoCopy function for that type (because they might not work and we only want
that to be an issue if an autoCopy is needed), but it seems that we don't
know how to respond if the autoCopy map doesn't have an entry for that type
because we had skipped it.

Add a test to cover that use case in a simpler form.