Comment out the super.init() in BaseClass's initializer
Both Brad and David felt that the body should not explicitly have a super.init()
in it.  Are there enough cases like this test that imply Phase 1 should be the
default?  Leave that question in the test (I won't do that for every test in
this category, I am tracking them elsewhere)