Add a future with Brad's request for the omitted initialization insertion point
Brad thinks we should have omitted initialization be inserted right after the
last explicit initialization, rather than just before the next or the
`super.init()` call (as it is today).  This would enable loops like the one in
this test to be Phase 1 operations, as the initialization would occur prior to
their operation.