Change isBetterMatch to a comparison routine
isBetterMatch actually already does most of the work of computing
if f1 is a better match than f2 *and* if f2 is a better match than f1.
This commit changes isBetterMatch to compareSpecificity which
returns -1 to prefer f1, 0 if they are equally preferred, and 1 to prefer
f2. Using this information we can correctly compute the set of
most specific functions, which could in some cases include the
wrong elements (and that is what is going wrong in issue #6578).

While there, I noticed that isBetterMatch/compareSpecificity was
not correctly handling the case when both functions contained
a most specific argument. In that case, the specificity is
ambiguous (according to the language specification) and so the
scope should be considered. See
  test/functions/ferguson/most-specific.chpl
which would have been an ambiguity error before this change
for an example.