Remove test/parallel/taskPar/rachels/Barrier.chpl
When looking through tests that used constructors that should be updated to
initializers, I ran across this test, which only defined a constant and a
record named Barrier.  It didn't seem to be used anywhere else and wasn't set
up to actually create an instance of the record, so I think it should be
removed.