Add tests to cover an issue with initializers and enums found by Sam
Exercises using enums in fields both with the enum type prefix and with a use
statement to allow unprefixed access.  Also shows what happens when you
incorrectly try to perform unprefixed access when a 'use' has not occurred -
normally, this gives you an error during function resolution (because it looks
like it might be a method) but it is tripping our initializers implementation
up and never reaching that nice error message.