Update test/types/records/hilde/newUserDefaultCtor.chpl for initializers
Convert the constructor into an initializer.  Due to normalization's handling
instance initialization, this no longer results in an initialization for the
type declaration and an initialization using the init expression and instead
gives the instance its type at normalization and then resolves the 'new' call
(meaning that the initializer is only called once whereas a constructor would
be called twice).  This meant that the expected output needed to change for the
removal of the temporary.