Update classes/deitz/constructors/test_generic_construct1 for initializers
Convert the constructor into an initializer.  Since type fields are not set
via the argument list anymore, the conversion required adding an initialization
of the type field in Phase 1 of the body.  I modified the name of the type
argument to avoid (my own) confusion between the argument and the field use for
the type of the concrete argument.