Clarify which field is generating type mismatch error
I ran into this error today, but couldn't figure out which field was
being complained about without having the error message name the
field.  This is a simple change that doesn't seem to affect any
existing tests unfortunately.  Also, I had trouble creating a test
that reproduces the error message (and didn't want to file the case
that originally caused me to run into it because I don't believe it
should be generating the error).