Add FLAG_COMPILER_ADDED_WHERE, use it in resolution
Resolution disambiguates functions based upon whether or not
they have a `where` clause. But, in certain situations,
the normalizer will add a `where` clause to a function
even if one did not exist before.

Since the `where` clause was added by the compiler, it's
not reasonable to expect Chapel programmers to understand
the implications on resolving otherwise ambiguous functions.

This commit
 * adds FLAG_COMPILER_ADDED_WHERE, sets it in the relevant
   part of the normalizer
 * checks FLAG_COMPILER_ADDED_WHERE in the resolution
   code that prefers functions using where clauses, to
   discount compiler-added where clauses
 * adds a test that would fail before this change