Add a bunch of tests for error messages, etc. in generic initializers
I realized that initializers on generic and concrete types were following
different code paths, so a lot of the error message improvements in the last 4
months only helped concrete types, not both as I had previously thought.  This
adds a bunch of tests to verify the behavior on generic types.  Not all of them
behave as desired today.