Allow UnresolvedSymExprs through the normalization of initializers
Prior to this change, if there was an unresolved symbol, we would just abort
with an internal error when encountering the AST node for it during the handling
of initializers.  This meant that code which would have later given a reasonable
error message instead didn't give a helpful one.  I updated the handling of
these AST nodes to allow them through so that function resolution would give us
the appropriate error message.