Relax filter of which functions will be stacktraced
For some time now, I've been feeling as though there are functions
missing from the CHPL_UNWIND output that shouldn't be -- most notably,
top-level user code like module initialization code and main().  It
turns out that this was due to the filter used in codegen to limit
which functions will be traceable and which not was too strict, based
on the C identifier having the prefix of 'chpl_' which eliminates such
routines.  Here, I apply the same test to the 'name' field such that
anything we munge with 'chpl' to avoid conflicts with C (like 'main')
will still be reported.  This improves the output significantly for
me.  It may be that even this test is a bit silly (e.g., maybe there's
a flag that we should be cueing off of), but this was a simple fix
that will make my use of CHPL_UNWIND much better, so it seemed
worthwhile on its own without getting distracted by trying to find the
optimal solution right now.