Include module 'init' routines in stack traces as well
This helps by getting the backtrace all the way up to the top
of the user's program.  The output could be made prettier by
avoding the extra obfuscation of the module initializer routine
names, but that's a bigger change than I want to take on here.
Also, this is mostly useful in the event that the stacktrace
lists the actual line numbers rather than the ones that declare
the procedure being reported on (since init routines will tend to
be line 1 in a file or the module delaration line), but this still
seems preferable to not reporting the top-level code point (and
Michael says that in some cases, we do print more precise line
numbers, though I have yet to see them).