Extend parallel array initialization to POD arrays
Ideally we want to be able to do parallel initialization for arrays of all
types, but we're currently blocked by an issue with arrays-of-arrays. The
problem is basically that an array's domain stores a linked list of all its
arrays and removal becomes expensive when addition and removal occur in
different orders. Long term we probably want to store the domain's arrays as an
associative domain or some data structure with < log(n) find/add/remove times.

We used to limit parallel array initialization to numeric types to workaround
the array-of-arrays issue. This just extends parallel array initialization to
support arrays of POD types as well.