Replace paratest.chapcs with an equivalent python version
paratest.chapcs was originally cobbled together from bash snippets that I used,
so it was written in bash. To fix some bugs in it I need to do some things that
would be hard/ugly to do in bash, so I rewrote the script in python in #6126.
This just replaces paratest.chapcs with the paratest.chapcs.py version. Note
that they should behave identically (bugs included) and I'll work on fixing the
bugs next.