Allow deadBlockElimination to remove unreachable do-while loops
PR #6631 fixed a bug in deadBlockElimination by preventing deadBlockElimination
from removing do-while loops just because the block containing the condition
variable was dead. This effectively meant that do-while loops could never be
dead block eliminated even if the entire loop was unreacahble. e.g.

    proc bar() {
      return 0;
      do { writeln("in loop"); } while true;
    }
    bar();

would leave an empty do-while loop. This updates deadBlockElimination to allow
removing the entire do-while loop if it's empty. The body of the loop (the
writeln above) is removed by dead block elimination since the body is
unreachable, which leaves the loop empty and allows it to be removed.