Update deadVariableElimination to revisit Symbols with removed uses
deadVariableElimination could be less effective depending on the order
in which Symobls were visted. Consider the following:

A = 1 + 3;
B = A * 5;

If 'A' were visted first, we would consider it a live variable because
of the single use in 'B'. If 'B' is visited first and found to be dead,
the single use of 'A' goes away and 'A' is now dead.

To accomplish this we fill a set with symbols used in statements that
are deleted as a variable is eliminated. If this symbol is not in the
current queue of symbols, we push it onto the queue.

This work exposed some compiler-introduced dead code in parallel and
buildOnStmt that is easily removed.