Fix the issue for newInit-*-record.chpl with _defaultOf calls
Basically, we weren't getting a good starting type for the field that was
declared to be an instance of a concrete record with initializers, though
concrete initializers weren't encountering this problem.  This was because
preNormalizeFields was only getting applied to fields on concrete types, not
generic types.  It isn't set up to handle param or type fields, understandably,
so make preNormalizeFields explicitly skip handling them.