Remove '=>' from additional tests
Mike found some more tests that were still using the '=>' operator which
is deprecated.  This rewrites them to use `ref` instead.

I completely removed the 'test_illegal_alias_operator' test because it
was designed to make sure that the alias operator couldn't be used as
a standalone statement, but that makes less sense if the operator
doesn't exist anymore (at least, in this context).

I also added a .bad file for 'desirable_syntax' which is a future in order
to improve our chances of tracking it as we go.