Explicitly specify that genericFieldInitTypeWoutInit will only handle params
This ought to be trivially true; type fields are only given an init statement,
and generic var/const by definition have no starting type or init information.

This makes the function assert that only params have been sent to it, and then
removes the checks around adding the flag to the temporaries we create.