Use the Barrier module in SSCA2 instead of defining its own barrier
SSCA2 has its own Barrier class that is very similar to the Barrier record
in the standard modules.  Delete that class and use the one in the modules
instead.

The barrier it was using was stored in a "Task Private Variables" data
structure, but after switching to the Barrier record, it was calling a
NULL virtual method.  Instead, just declare a barrier variable right before
it is used.  Also, pass the barrier to a coforall with 'ref' intent.