Update these .good files to match current behavior
Here's what's going on: The range 1..4 is reindexed to 2..9 by 2.  We
then make a 'local subdoain' query on the reindexed array which passes
the query down to the underlying domain.  The underlying domain says
"I own 1..4" which we then translate up to 2..8 [by 2] in the
upward-facing index set.  This is equivalent to 2..9 by 2 but
obviously not identical.  That said, I don't think there's any general
way that we can/should change 2..8 by 2 into 2..9 by 2.  So I'm
updating the .good files instead.