Turn off affinity for whitebox testing to (hopefully) avoid timeouts
We've been seeing sporadic timeouts for whitebox testing since mid May. For
most of the timeouts it looks like 2 executables were running at the same time,
and we know that multiple qthreads instances don't play well together. We try
to limit to 1 running executable with CHPL_TEST_LIMIT_RUNNING_EXECUTABLES, but
that uses a file as a lock, which can be racey.

To hopefully avoid timeouts when that race does occur, disable qthreads
affinity and limit the amount of spinwaiting qthreads does. Note that we don't
disable affinity for pgi, because pgi doesn't support atomics. This forces
qthreads to use mutexes instead of spinlocks, which when combined with
disabling affinity has lead to timeouts in the past.

Note that this is similar to previous attempts like #4060 and #3684, except
that we don't disable affinity for pgi. Also note that back then multiple
executables were running together all the time since we weren't using
CHPL_TEST_LIMIT_RUNNING_EXECUTABLES, but now that should only happen on rare
occasions.