Update one of the helper modules for test/studies/madness/aniruddha/madchap
Converts a constructor into an initializer.

Currently encounters an assert in initializerRules.cpp because of the coforall
on structure (I'm pretty sure).  Open question as to whether this is the correct
modification of the initializer.

Does not update MRA.chpl (the other helper module) as that initializer is huge
and runs into a segfault during copyPropagation when I tried the obvious
conversion.