Update test/types/records/hilde/callConstructorThruTypeFn to use initializers
Replace the constructors in this test with initializers.  R could benefit from
Phase 1 as the default.

This test fails today with the message "illegal use of function that does not
return a value: 'init'".  I believe this is because the type function is
causing the 'new' call to be transformed during function resolution but that
we don't expect that to happen to non-generic records with initializers at the
moment.