Convert test/types/records/hilde/callConstructorThruTypeVar for initializers
Replace the constructor with an initializer.  Note that this initializer could
benefit from Phase 1 being the default.

Currently fails for the same reason that callConstructorThruTypeFn fails, which
is that the type variable is causing the 'new' call to be transformed during
function resolution when we expect non-generic records with initializers to
handle this solely during normalization.