Update studies/matrixMult/stonea/distributed/simpleBlockDist for initializers
Convert the constructors into initializers.  The non-trivial initializer would
likely benefit from Phase 1 as the default.

This test currently runs into problems because we don't expect non-generic
records with initializers to need a "_defaultOf" function, but when the type is
passed in as an argument to a function and that argument is used to declare an
instance, PRIM_INIT and thus a _defaultOf call are still inserted.