Remove initializer from *atomics version; simplify field inits
I couldn't get the atomics version to work simply by changing the
constructor to an initializer because of the issue in
initAtomicsViaWrite.chpl.  But I also realized that the initializer
isn't really necessary since 0 is the guaranteed default value of the
atomic ints anyway.  Rewrote to rely on that, commenting on it,
and combining the two fields into one; then did a similar combination
for the *sync version.