Update .good to match expected output
Thinking about this with a clear head, I'm realizing that the new
interpretation is that _defaultOf() is ignored -- just like any other
function -- and that init() wins.  I believe that this is the correct
behavior in the new world -- a 0-argument initializer will be the new
way to specify a default value.