Add four new tests of coforalls in initializers
Covers coforalls that contain `this.init()`, `super.init()` and Phase 1 field
initializations, as well as a coforall in Phase 1 that does not contain a field
init.

These are all for initializers on concrete types, but I have no reason to
suspect initializers on generic types are not appropriately supported.

These tests all failed before the change I will in this branch.