Remove handling of references from copyPropagation
As a part of our goal to simplify and/or rework copyPropagation, this
commit removes the parts of the copyPropagation pass that attempt to
propagate references. Our alias analysis capabilities are poor and the
benefit of this code is tiny compared to the effort required to maintain
it. After removing this code, miniMD only suffers a .01% increase in
AST size.

While here I made some minor style cleanups:
- used a printf macro instead of checking against DEBUG_CP repeatedly
- cache the result of 'call->get(1) == se'