Fix the record type alias tests that fail.
Normalization tries to replace "new" calls to records with initializers with
just calls to the initializer in the right format.  Unfortunately this isn't
always possible, such as in the case of using type aliases.  This change allows
records that we find through type aliases to appropriately replace PRIM_NEW
at resolution (instead of normalization, which isn't possible).