Fix bug where variable was propagated after it was redefined
copyPropagation was propagating A in AST like this:
  (move A B)
  (move ref_tmp (addrof B))
  (+= B 1)
  (call writeln A)
such that the writeln became:
  (call writeln B)

The problem was that when B's address was taken copyPropagation should
have killed any alias/def pairs with B on the RHS. I realized that the
existing addr-of/liveRefs logic could move into 'isDef' where it would
then trigger a call to ``removeAvailable``.

With this change the ``liveRefs`` set is no longer necessary.