Squash some reference-related bugs
It turns out the 'liveRefs' capability was useful for fixing some
existing bugs. The basic idea is that we should never create a new copy
pair if either side has been referenced.

Also updates 'isUse' to check if a record is passed by const-ref intent,
a situation in which a const-ref to a field can be returned. Replacing
the record can result in a reference to different data.