Add FLAG_LAST_RESORT, use it
FLAG_COMPILER_GENERATED previously included multiple overloaded meanings.
The most confusing of these overloaded meanings was that a function
marked with FLAG_COMPILER_GENERATED is only selected during resolution if
candidate functions without that flag are found. This was sometimes used
to create a last-resort function for errors (such as with deprecation
errors) even though the function involved is not compiler generated (or
even internal to the implementation).

This commit creates a new flag, FLAG_LAST_RESORT, to encode specifically
this behavior in function resolution.

To implement the change, I looked at every use of FLAG_COMPILER_GENERATED
in the compiler and modules.

When FLAG_COMPILER_GENERATED was added to a function, I generally
also added FLAG_LAST_RESORT. In some cases I was confident that the
intent was *only* to make the function last-resort in resolution (as with
cases where the function exists to produce an error message). In those
cases I removed the FLAG_COMPILER_GENERATED function.

When FLAG_COMPILER_GENERATED was checked, I verified that its use made
sense - e.g. that it is controlling error output or had to do with
identifying POD types. I updated the comment in flags_list.h for
FLAG_COMPILER_GENERATED to reflect my understanding of what impact
this flag has.

Of course I updated the resolution portion that preferred functions
without FLAG_COMPILER_GENERATED to instead check for FLAG_LAST_RESORT.