Add warning message for a case of module name sharing
If there is no module that encloses all the executable statements in the file,
then we create one with the file name.  This could be confusing if a user
thought their module with the same name as the file was at the top level, when
really it wasn't, such as here:

var x: int;
module SameName {
  var y = 21;
}

This change adds a warning message when the module you have defined has a parent
with the same name and both have the same name as the file.