Update gasnet XE/XC cross configure scripts
The XE/XC cross-configure scripts were renamed in the 1.28.2 release, so update
our references to them accordingly. Also, it appears that gasnet has dropped
the cross-configure script for XT. We dropped module support for XT some time
ago, but I wonder if it's time to drop from-source support as well.