Cleaning up some futures
De-futurizing:

  # due probably to #6722
  distributions/kushal/assignDomain

  # due probably to #6765
  parallel/begin/vass/begin-in-module-init
  users/ferguson/bug-zhao

While there, removed a .prediff as it should not be there.

While there also: converted the following test
so that it always compiles with --verify.
Compiler's crashing under --verify  is a more immediate
and more reliable bug than compiling the program
all the way (which can be done without --verify).

  types/records/engin/parentheseslessConstErrorField