Add an error message for record inheritance
We removed record inheritance from the language specification due to not having
ironed out enough of the details.  However, field inheritance for records was
already implemented, and we did not remove it.

This change adds an error message after parsing if we detect that a record has
been defined which inherits.  It also points users encountering this to an
issue I created for discussion on record inheritance.

It does not remove field inheritance for records.