Modify error message when both 'require' and 'use' statements are present
Michael asked me to write a test for when both of these statements are present
(but no others are) and cause the warning, which made me realize that it would
be annoying to only call out one type and not both when both were present.  So
modify the error message in the case where both are present.