Test changes for the error message on record inheritance
Make several tests of record inheritance into futures, and update the expected
failure mode for several futures dealing with record inheritance for the new
error message.

This should resolve most of the failures detected in this branch, with the
exception of test/users/vass/type-tests.isSubtype.chpl (which I will
intentionally handle in a separate commit).