Make this initializer future more robust to module changes
Because it fails to resolve the init call, we get a bunch of output from the
standard modules.  Fixing that additional output is a separate issue, but the
output is likely to change as modules get updated - add a prediff to filter out
the module failures so that the output is deterministic to any module changes
that occur (and update the expected .good and .bad output accordingly).