Renamed Toml.chpl to TOML.chpl
Toml.chpl was introduced in #6752
Judging by the error message in this file:

    test/modules/packages/Toml/test/quoteKeyPeriod.bad

it was tested as TOML.chpl. I am renaming it so that the tests pass.

I am also adding TOML.notest because Toml.chpl lacks a .good file
and is probably not intended to be tested by itself anyway.