Fix LLVM testing failures from PR #6856
type.h previously included this line,

  #define LOCALE_ID_TYPE dtLocaleID->typeInfo()

which was only used for LLVM code generation. So I replaced one use of
this macro in codegen/type.cpp with its definition, and I moved the
define to codegen/expr.cpp where most of the uses are.

I would expect we could entirely remove LOCALE_ID_TYPE and replace it
with just dtLocaleID but here I'm making a more conservative change since
my goal is simply to repair testing.