Separate memory registration information from comm domains.
We have been registering memory separately on each comm domain, and
storing the resulting GNI memory descriptor handles in our comm domain
description structs.  But when we look up the local memory registration
we don't look at the memory table on the CD we will be using, because we
haven't allocated one yet.  Instead we just look at the table on CD 0.
(We used to allocate a CD before looking up the local registration, but
long ago we implemented a change that delayed CD allocation as long as
possible, without doing anything about the memory registration.)

Here, pull the memory registration table out of the comm domain struct
into a static array in memory.  This could evolve into a more complex
data structure in the future, but this is good enough for now.

The initial version of this produced a significant performance drop (35%
or so) for ra-atomics on XC-16.  It turned out that the memory table had
been serving an unrecognized role in the comm domain struct, making the
comm_doms[] elements large enough to prevent any cache line sharing
across them.  With the registration table gone, two threads holding
successive comm_doms[] elements could create cache conflicts for each
other.  The new cache_spacer element of comm_dom_t prevents this now.

While here, I also did several minor cleanups in the memory registration
logic and moved the registration lookup functions next to the ones that
set up the table.