Add another test of type fields and default values for type arguments
I got myself confused when looking at the test/studies/hpcc/ changes and
convinced myself there was something wrong with default values for type
arguments when the type field being assigned to was declared with an initial
value.  Turns out the problem was I had only partially converted the constructor
into an initializer (having forgotten to update the name).  But I wrote this
test when investigating and it worked just fine so might as well keep it.