Convert type-based alias analysis to LLVM 4.0 format (#6875)
Convert type-based alias analysis to LLVM 4.0 format (#6875)

[reviewed by @mppf ]

This converts from the old TBAA to the new struct-path TBAA format for LLVM.

There used to be just one kind of TBAA node for normal accesses. Now there are two: the type descriptor and the access tag. The access tag points to the type descriptor, and `!tbaa` annotations should always use the access tag.

This change only converts the format. In the long run, handling of struct members needs improvement. They are correct, just suboptimal because right now they can alias with anything of the member type.